reduce AbstractBaseComponentTest
