+++ celements-base-test/src/main/java/com/celements/common/test/AbstractBaseComponentTest.java
@@ -1,17 +1,11 @@
 package com.celements.common.test;
 
-import java.io.File;
-import java.io.InputStream;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.util.List;
 
 import org.junit.After;
 import org.junit.Before;
 import org.xwiki.component.manager.ComponentManager;
 import org.xwiki.component.spring.XWikiSpringConfig;
-import org.xwiki.container.ApplicationContext;
-import org.xwiki.container.Container;
 
 import com.celements.spring.CelSpringConfig;
 import com.celements.spring.context.CelSpringContext;
@@ -20,11 +14,9 @@ import com.google.common.collect.ImmutableList;
 public abstract class AbstractBaseComponentTest {
 
   @Before
-  public void setUp() throws Exception {
+  public void setUpSpringContext() throws Exception {
     CelSpringContext springCtx = new CelSpringContext(getSpringConfigs());
     CelementsSpringTestUtil.setContext(springCtx);
-    springCtx.getBean(Container.class) // TODO do this in lower class?
-        .setApplicationContext(new TestXWikiApplicationContext());
   }
 
   /**
@@ -35,7 +27,7 @@ public abstract class AbstractBaseComponentTest {
   }
 
   @After
-  public void tearDown() throws Exception {
+  public void tearDownSpringContext() throws Exception {
     CelementsSpringTestUtil.removeContext()
         .close();
   }
@@ -46,27 +38,6 @@ public abstract class AbstractBaseComponentTest {
 
   public ComponentManager getComponentManager() {
     return getSpringContext().getBean(ComponentManager.class);
-  }
-
-  public static class TestXWikiApplicationContext implements ApplicationContext {
-
-    @Override
-    public URL getResource(String resourceName) throws MalformedURLException {
-      if (resourceName.contains("xwiki.properties")) {
-        return this.getClass().getClassLoader().getResource("xwiki.properties");
-      }
-      throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public InputStream getResourceAsStream(String resourceName) {
-      throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public File getTemporaryDirectory() {
-      throw new UnsupportedOperationException();
-    }
   }
 
 }
